feat: use a compact string for label names and values
Label names and values are commonly small strings, which can be stored
efficiently on the stack if they are below a certain size. This module
provides the `CompactString` type which makes use of this optimization
to avoid having to heap allocate in the common case.

The current implementation effectively reuses
`compact_str::CompactString`. We use a newtype wrapper so we can
implement the required pyo3 traits and to make it easier to modify
later if we need to.

This gets us a decent speedup in the benchmark:

```
❯ cargo bench
   ...
     Running benches/my_benchmark.rs (target/release/deps/my_benchmark-81bd16a5775629d5)
merge_internal          time:   [75.765 µs 75.840 µs 75.922 µs]
                        change: [-23.539% -23.365% -23.161%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe
```